Move betty.plugins.link to betty.links
